+++ src/Maritaria/CheatBlocks/InfiniteFuelBlock.cs
