+++ src/Maritaria/CheatBlocks/ModuleInfiniteFuel.cs
