+++ src/Maritaria/Cursor/CursorManager.cs
