+++ src/Maritaria/Cursor/CursorType.cs
