+++ src/Maritaria/Cursor/GameCursor.cs
