+++ src/Maritaria/MagnetToggleKeyBehaviour.cs
