+++ src/Maritaria/Maritaria.csproj
@@ -4,19 +4,15 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{129AE68B-6968-44C4-A386-68C0E1E9A63E}</ProjectGuid>
+    <ProjectGuid>{8831864E-3AEE-4B9B-B82C-0DC0172D5626}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Nuterra</RootNamespace>
-    <AssemblyName>Nuterra</AssemblyName>
-    <TargetFrameworkVersion>v3.5</TargetFrameworkVersion>
+    <RootNamespace>Maritaria</RootNamespace>
+    <AssemblyName>Maritaria</AssemblyName>
+    <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
-    <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
-    <TargetFrameworkProfile>
-    </TargetFrameworkProfile>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <PlatformTarget>AnyCPU</PlatformTarget>
     <DebugSymbols>true</DebugSymbols>
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
@@ -26,7 +22,6 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <PlatformTarget>AnyCPU</PlatformTarget>
     <DebugType>pdbonly</DebugType>
     <Optimize>true</Optimize>
     <OutputPath>bin\Release\</OutputPath>
@@ -39,33 +34,44 @@
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
+    <Reference Include="Microsoft.CSharp" />
     <Reference Include="System.Data" />
+    <Reference Include="System.Net.Http" />
     <Reference Include="System.Xml" />
-    <Reference Include="Newtonsoft.Json, Version=6.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed">
-      <HintPath>..\..\lib\Newtonsoft.Json.dll</HintPath>
-    </Reference>
-    <Reference Include="UnityEngine, Version=0.0.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <SpecificVersion>False</SpecificVersion>
-      <HintPath>..\..\lib\UnityEngine.dll</HintPath>
-    </Reference>
-    <Reference Include="UnityEngine.UI">
-      <HintPath>..\..\lib\UnityEngine.UI.dll</HintPath>
-    </Reference>
-    <Reference Include="Assembly-CSharp">
-      <SpecificVersion>False</SpecificVersion>
-      <HintPath>..\..\lib\Assembly-CSharp.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="BaconBlock.cs" />
+    <Compile Include="BlockLoader.cs" />
+    <Compile Include="BugReportFlagger.cs" />
+    <Compile Include="CheatBlocks\CheatBlockMod.cs" />
+    <Compile Include="CheatBlocks\GodModeBlock.cs" />
+    <Compile Include="CheatBlocks\InfiniteFuelBlock.cs" />
+    <Compile Include="CheatBlocks\ModuleGodMode.cs" />
+    <Compile Include="CheatBlocks\ModuleInfiniteFuel.cs" />
+    <Compile Include="Cursor\CursorManager.cs" />
+    <Compile Include="Cursor\CursorType.cs" />
+    <Compile Include="Cursor\CustomCursorMod.cs" />
+    <Compile Include="Cursor\GameCursor.cs" />
+    <Compile Include="CustomBlock.cs" />
+    <Compile Include="FirstPerson\CockpitBlock.cs" />
+    <Compile Include="FirstPerson\FirstPersonCamera.cs" />
+    <Compile Include="FirstPerson\FirstPersonController.cs" />
+    <Compile Include="FirstPerson\LargeCockpitBlock.cs" />
+    <Compile Include="FirstPerson\ModuleFirstPerson.cs" />
+    <Compile Include="GameObjectExtensions.cs" />
+    <Compile Include="MagnetToggleKeyBehaviour.cs" />
+    <Compile Include="MaritariaConfig.cs" />
+    <Compile Include="MaritariaMod.cs" />
+    <Compile Include="Modules.cs" />
+    <Compile Include="ProductionStoppedSign.cs" />
+    <Compile Include="ProductionToggleKeyBehaviour.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="AssetBundleImport.cs" />
-    <Compile Include="CleanLogger.cs" />
-    <Compile Include="ModAttribute.cs" />
-    <Compile Include="ModConfig.cs" />
-    <Compile Include="ModLoader.cs" />
-    <Compile Include="Nuterra.cs" />
-    <Compile Include="TerraTechMod.cs" />
+    <Compile Include="SaveGameFlagger.cs" />
+    <Compile Include="SplashScreenHandler.cs" />
+    <Compile Include="SpriteFactory.cs" />
+    <Compile Include="SuicideKeyBehaviour.cs" />
+    <Compile Include="TimeOfDayKeysBehaviour.cs" />
+    <Compile Include="UnityGraph.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
