+++ src/Maritaria/ProductionStoppedSign.cs
