+++ src/Maritaria/SplashScreenHandler.cs
