+++ src/Maritaria/SpriteFactory.cs
