+++ src/Maritaria/SuicideKeyBehaviour.cs
