+++ src/Nuterra.sln
@@ -13,6 +13,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{CC96E61F-7
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nuterra.Internal", "Nuterra.Internal\Nuterra.Internal.csproj", "{306B93ED-DC73-4899-BB32-D9571152C554}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Mods", "Mods", "{3BA7DE14-6278-440A-95B1-026F30C10140}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Maritaria", "Maritaria\Maritaria.csproj", "{8831864E-3AEE-4B9B-B82C-0DC0172D5626}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sylver", "Sylver\Sylver.csproj", "{2E708037-AD6B-49DB-9B7D-AF459E69E386}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -35,6 +41,14 @@ Global
 		{306B93ED-DC73-4899-BB32-D9571152C554}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{306B93ED-DC73-4899-BB32-D9571152C554}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{306B93ED-DC73-4899-BB32-D9571152C554}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8831864E-3AEE-4B9B-B82C-0DC0172D5626}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8831864E-3AEE-4B9B-B82C-0DC0172D5626}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8831864E-3AEE-4B9B-B82C-0DC0172D5626}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8831864E-3AEE-4B9B-B82C-0DC0172D5626}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2E708037-AD6B-49DB-9B7D-AF459E69E386}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2E708037-AD6B-49DB-9B7D-AF459E69E386}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2E708037-AD6B-49DB-9B7D-AF459E69E386}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2E708037-AD6B-49DB-9B7D-AF459E69E386}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -44,5 +58,7 @@ Global
 		{CFB5CD3A-47A1-4E71-BA81-C93D4BB657B3} = {CC96E61F-75DD-4644-AC51-D9B783C00680}
 		{F953D402-AAE3-4335-90A6-13BDBDF6D4FA} = {CC96E61F-75DD-4644-AC51-D9B783C00680}
 		{306B93ED-DC73-4899-BB32-D9571152C554} = {CC96E61F-75DD-4644-AC51-D9B783C00680}
+		{8831864E-3AEE-4B9B-B82C-0DC0172D5626} = {3BA7DE14-6278-440A-95B1-026F30C10140}
+		{2E708037-AD6B-49DB-9B7D-AF459E69E386} = {3BA7DE14-6278-440A-95B1-026F30C10140}
 	EndGlobalSection
 EndGlobal
